requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.27
langchain==0.1.9
langchain-openai==0.0.5
openai==1.12.0
pymupdf==1.23.21
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2
python-pptx==0.6.22
weaviate-client==3.25.3
supabase==2.3.1
python-dotenv==1.0.1
boto3==1.34.34
pdf2image==1.17.0
unstructured==0.17.2
requests==2.31.0
